+++ src/test/java/de/alpharogroup/collections/pairs/ImmutableBoxTest.java
@@ -30,9 +30,9 @@ import static org.testng.AssertJUnit.assertTrue;
 
 import org.testng.annotations.Test;
 
-import de.alpharogroup.evaluate.object.EqualsEvaluator;
-import de.alpharogroup.evaluate.object.HashcodeEvaluator;
-import de.alpharogroup.evaluate.object.ToStringEvaluator;
+import de.alpharogroup.evaluate.object.evaluators.EqualsEvaluator;
+import de.alpharogroup.evaluate.object.evaluators.HashcodeEvaluator;
+import de.alpharogroup.evaluate.object.evaluators.ToStringEvaluator;
 
 /**
  * The unit test class for the class {@link ImmutableBox}.