+++ src/main/java/de/alpharogroup/collections/list/ListFactory.java
@@ -81,8 +81,12 @@ public final class ListFactory
 	 *            the generic type of the elements
 	 * @param initialCapacity
 	 *            the initial capacity
-	 * @return the new {@link ArrayList} as {@link List}.
+	 * @return the new {@link ArrayList} as {@link List}
+	 * @deprecated use instead the same name method with elements <br>
+	 *             <br>
+	 *             Note: will be removed in next minor release
 	 */
+	@Deprecated
 	public static <T> List<T> newArrayList(final int initialCapacity)
 	{
 		return new ArrayList<>(initialCapacity);