+++ src/test/java/de/alpharogroup/collections/list/ListFactoryTest.java
@@ -61,6 +61,7 @@ public class ListFactoryTest
 	/**
 	 * Test the method {@link ListFactory#newArrayList(int)}.
 	 */
+	@SuppressWarnings("deprecation")
 	@Test
 	public void testNewArrayListInt()
 	{
@@ -85,7 +86,6 @@ public class ListFactoryTest
 	/**
 	 * Test the method {@link ListFactory#newRangeList(int, int)}
 	 */
-	@SuppressWarnings("serial")
 	@Test
 	public void testNewRangeList()
 	{